minor aesthetic changes
